feat: add types in previsions of NPC travels (#50)
* Add types in previsions of NPC travels

* change timestamps types to u64

* Change use of uses and .vscode/settings.json

* Change contract components

* change current time to u64 in can_spawn

* Change public key + format

* Fixed tests

---------

+++ contracts/src/models/config.cairo
@@ -198,7 +198,7 @@ struct LevelingConfig {
 struct NpcConfig {
     #[key]
     config_id: u128,
-    spawn_delay: u128,
+    spawn_delay: u64,
     pub_key: felt252
 }
